requirements.txt
torch
torchvision
torchaudio
hydra-core
neptune-client
tensorboard
torchmetrics